Refactor: @emotion/styled to Popup.jsx
