+++ babel.config.js
@@ -1,0 +1,2 @@
+export const presets = ['@babel/preset-env', '@babel/preset-react'];
+export const plugins = ['@emotion'];
